[IAP]Refactor the in-app purchase of Google Play
This patch refactors the previous implementation of in-app purchase(IAP)
on Android. The main differences are:

1. Redefine the javascript interfaces to combine with the one used on
iOS, thus developers can deploy the same code on both platforms. The
spec locates at https://github.com/crosswalk-project/ios-extensions-crosswalk
/blob/master/extensions/InAppPurchase/spec/iap.html.

2. Refactor the framework of the IAP extension in order to add
additional stores later. The extension itself acts as a manager of
different backends, it forwards the commands to the proper backend if it
supports. To add support of new backend, just inherit the
class InAppPurchaseHelper and implement the command methods.

3. Refactor the implementation of Google Play to reject the promise when
an error happens.

BUG=XWALK-5696